found some bugs (with dmd/std)
found some bugs (with multi_index)
 retarded things I did:
  1: got the index into the hash table, then perform a rehash
  2: forgot the mod hashlen
  3: off by 1 error in heap parent function
  4: didn't read boost::multi_index' future work closely enough. oops.
but finished signals and slots thing anyways. The way I have it set up
is {Value type} is the signal, and {multi_index container} is the slot.
What Muñoz proposed was {multi_index container} is the signal, and the
slot is totally user defined, which is probably more valuable.

Concerning the rest of his future work: ranked indeces might be interesting,
constraints sound complicated and will need some actual discussion with the guy,
user defined indeces are trivial, indexed maps sound kind of redundant, and
move semantics come for free in D (afaik).